+++ {{cookiecutter.project_slug}}/{{cookiecutter.package_name}}/developmentsettings.py
@@ -1,4 +1,4 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
+# (c) Nelen & Schuurmans.  Proprietary, see LICENSE file.
 from {{ cookiecutter.package_name }}.settings import *
 
 DEBUG = True
@@ -6,31 +6,16 @@ DEBUG = True
 LOGGING['loggers']['']['level'] = 'DEBUG'
 RAVEN_CONFIG = {}
 
-# ENGINE: 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
-# In case of geodatabase, prepend with:
-# django.contrib.gis.db.backends.(postgis)
 DATABASES = {
-    # If you want to use another database, consider putting the database
-    # settings in localsettings.py. Otherwise, if you change the settings in
-    # the current file and commit them to the repository, other developers will
-    # also use these settings whether they have that database or not.
-    # One of those other developers is Jenkins, our continuous integration
-    # solution. Jenkins can only run the tests of the current application when
-    # the specified database exists. When the tests cannot run, Jenkins sees
-    # that as an error.
     'default': {
-        'NAME': os.path.join(BUILDOUT_DIR, 'var', 'sqlite', 'test.db'),
-        'ENGINE': 'django.contrib.gis.db.backends.spatialite',
-        # If you want to use postgres, use the two lines below.
-        # 'NAME': '{{ cookiecutter.package_name }}',
-        # 'ENGINE': 'django.contrib.gis.db.backends.postgis',
+        'NAME': '{{ cookiecutter.package_name }}',
+        'ENGINE': 'django.db.backends.postgresql',
         'USER': 'buildout',
         'PASSWORD': 'buildout',
-        'HOST': '',  # empty string for localhost.
-        'PORT': '',  # empty string for default.
-        }
+        'HOST': 'db',  # 'db' is the postgres docker name
+        'PORT': 5432,
     }
-
+}
 
 try:
     from {{ cookiecutter.package_name }}.localsettings import *