+++ {{cookiecutter.project_slug}}/{{cookiecutter.package_name}}/tests.py
@@ -1,9 +1,13 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
+# (c) Nelen & Schuurmans.  Proprietary, see LICENSE file.
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 from __future__ import print_function
 
 from django.test import TestCase
+from {{ cookiecutter.package_name }} import admin
+from {{ cookiecutter.package_name }} import models
+from {{ cookiecutter.package_name }} import urls
+from {{ cookiecutter.package_name }} import views
 
 
 class ExampleTest(TestCase):