+++ {{cookiecutter.project_slug}}/{{cookiecutter.package_name}}/urls.py
@@ -1,31 +1,23 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
+# (c) Nelen & Schuurmans.  Proprietary, see LICENSE file.
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 from __future__ import print_function
 
 from django.conf.urls import include
-from django.conf.urls import patterns
 from django.conf.urls import url
 from django.contrib import admin
-from lizard_map.views import HomepageView
-from lizard_ui.urls import debugmode_urlpatterns
+from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 
 from {{ cookiecutter.package_name }} import views
 
-admin.autodiscover()
 
-urlpatterns = patterns(
-    '',
-    url(r'^$', HomepageView.as_view()),
+urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
-    url(r'^webmap/', include('lizard_wms.urls')),
-    url(r'^map/', include('lizard_map.urls')),
-    url(r'^ui/', include('lizard_ui.urls')),
     # url(r'^something/',
     #     views.some_method,
     #     name="name_it"),
     # url(r'^something_else/$',
     #     views.SomeClassBasedView.as_view(),
     #     name='name_it_too'),
-    )
-urlpatterns += debugmode_urlpatterns()
+]
+urlpatterns += staticfiles_urlpatterns()