+++ {{cookiecutter.project_slug}}/{{cookiecutter.package_name}}/views.py
@@ -1,23 +1,9 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
+# (c) Nelen & Schuurmans.  Proprietary, see LICENSE file.
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 from __future__ import print_function
 
 from django.utils.translation import ugettext as _
 # from django.core.urlresolvers import reverse
-# from lizard_map.views import MapView
-# from lizard_ui.views import UiView
 
 # from {{ cookiecutter.package_name }} import models
-
-
-# class TodoView(UiView):
-#     """Simple view without a map."""
-#     template_name = '{{ cookiecutter.package_name }}/todo.html'
-#     page_title = _('TODO view')
-
-
-# class Todo2View(MapView):
-#     """Simple view with a map."""
-#     template_name = '{{ cookiecutter.package_name }}/todo2.html'
-#     page_title = _('TODO 2 view')