+++ CHANGES.rst
@@ -1,9 +1,12 @@
 Changelog for cookiecutter-djangosite-template
 ==============================================
 
+
 0.1 (unreleased)
 ----------------
 
 - Started project.
 
 - Got first version of the template to work.
+
+- Added docker test setup, Jenkins file and ansible integration.